Add builder into its own file
