+++ packages/doenetml-worker-rust/lib-doenetml-core/src/new_core.rs
@@ -1,3 +1,4 @@
+pub mod component_builder;
 pub mod graph_based_core;
 pub mod graph_node;
 #[cfg(test)]
